% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sector-regional-nation.R
\name{io_sector_regional_nation}
\alias{io_sector_regional_nation}
\title{Get sector information for the national input-output tables}
\usage{
io_sector_regional_nation(year, axis)
}
\arguments{
\item{year}{Year of the data.}

\item{axis}{Axis of the data, either `"input"` or `"output"`.}
}
\value{
A data frame containing the sector information for the specified year and axis.
}
\description{
Get sector information for the national input-output tables
}
