% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/econiodatajp-package.R
\docType{package}
\name{econiodatajp-package}
\alias{econiodatajp}
\alias{econiodatajp-package}
\title{econiodatajp: Dataset for the input-output tables for Japan}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\author{
\strong{Maintainer}: Mizuki Uchida \email{uchidamizuki@vivaldi.net}

}
\keyword{internal}
