% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-get.R
\name{io_table_regional_get_nation}
\alias{io_table_regional_get_nation}
\title{Get the national input-output table}
\usage{
io_table_regional_get_nation(
  year,
  price_type = "producer_price",
  sector_class = c("basic", "small", "medium", "large", "template"),
  competitive_import = TRUE
)
}
\arguments{
\item{year}{Year of the data.}

\item{price_type}{Price basis of the table. Currently only
`"producer_price"` is available.}

\item{sector_class}{Sector classification granularity.}

\item{competitive_import}{Whether to use the competitive-import
convention (matches the `competitive_import` argument of
[econioread::io_table_read_sector_types()]). `"basic"` and `"small"`
sector classes are only available when `TRUE`.}
}
\value{
An input-output table object for the specified year.
}
\description{
Get the national input-output table
}
